Core/PacketIO: Fix packet structure for CMSG_SEND_PING_WORLD_POINT
Port From (https://github.com/TrinityCore/TrinityCore/commit/38e6171ec6d1acf0377c8f6e75f5f39d6701f3bf)

+++ Source/Game/Networking/Packets/PartyPackets.cs
@@ -1049,7 +1049,7 @@ namespace Game.Networking.Packets
             SenderGUID = _worldPacket.ReadPackedGuid();
             MapID = _worldPacket.ReadUInt32();
             Point = _worldPacket.ReadVector3();
-            Type = (PingSubjectType)_worldPacket.ReadUInt32();
+            Type = (PingSubjectType)_worldPacket.ReadInt8();
             PinFrameID = _worldPacket.ReadUInt32();
             Transport = _worldPacket.ReadPackedGuid();
             PingDuration = TimeSpan.FromMilliseconds(_worldPacket.ReadInt32());
